add site search and fixed migration file

+++ plugins/tai/products/updates/builder_table_create_tai_products_products_orders.php
@@ -13,7 +13,7 @@ class BuilderTableCreateTaiProductsProductsOrders extends Migration
             $table->integer('id');
             $table->integer('product_id');
             $table->integer('order_id');
-            $table->double('quantity', 10, 0);
+            $table->decimal('quantity', 10, 0);
         });
     }
